Fixed Linting & Updated Snapshots
